* Add RunBars class * Add InformationBars class * Add class inheritance * Add Docstrings * RM redundant imports
