+++ mlfinlab/data_structures/run_data_structures.py
@@ -3,83 +3,92 @@ Advances in Financial Machine Learning, Marcos Lopez de Prado
 Chapter 2: Financial Data Structures
 
 This module contains the functions to help users create structured financial data from raw unstructured data,
-in the form of tick, volume, and dollar imbalance bars.
+in the form of tick, volume, and dollar run bars.
 
-These bars are used throughout the text book (Advances in Financial Machine Learning, By Marcos Lopez de Prado, 2018, pg 25)
-to build the more interesting features for predicting financial time series data.
+These bars are used throughout the text book (Advances in Financial Machine Learning, By Marcos Lopez de Prado, 2018,
+pg 31) to build the more interesting features for predicting financial time series data.
 
-These financial data structures have better statistical properties when compared to those based on fixed time interval sampling.
-A great paper to read more about this is titled: The Volume Clock: Insights into the high frequency paradigm, Lopez de Prado, et al
+These financial data structures have better statistical properties when compared to those based on fixed time interval
+sampling. A great paper to read more about this is titled: The Volume Clock: Insights into the high frequency paradigm,
+Lopez de Prado, et al
 """
 
 # Imports
 from collections import namedtuple
-import pandas as pd
 import numpy as np
 from mlfinlab.util.fast_ewma import ewma
-
-
-class InformationBars:
-    def __init__(self, file_path, metric, exp_num_ticks_init=100000, num_prev_bars=3, num_ticks_ewma_window=20,
-                 batch_size=2e7):
-
-        # Base properties
-        self.file_path = file_path
-        self.metric = metric
-        self.exp_num_ticks_init = exp_num_ticks_init
-        self.num_prev_bars = num_prev_bars
-        self.num_ticks_ewma_window = num_ticks_ewma_window
-        self.batch_size = batch_size
-
-        # Batch_run properties
-        self.flag = False  # The first flag is false since the first batch doesn't use the cache
-        self.cache = []
-        self.num_ticks_bar = []  # List of number of ticks from previous bars
+from mlfinlab.data_structures.information_bars import InformationBars
 
 
 class RunBars(InformationBars):
-    def __init__(self, file_path, metric, exp_num_ticks_init=100000,
-                 num_prev_bars=3, num_ticks_ewma_window=20, batch_size=2e7):
+    """
+    Contains all of the logic to construct the run bars from chapter 2. This class shouldn't be used directly.
+    We have added functions to the package such as get_dollar_run_bars which will create an instance of this
+    class and then construct the run bars, to return to the user.
 
+    This is because we wanted to simplify the logic as much as possible, for the end user.
+    """
+
+    def __init__(self, file_path, metric, exp_num_ticks_init=100000, num_prev_bars=3, num_ticks_ewma_window=20,
+                 batch_size=2e7):
+        """
+        Constructor
+
+        :param file_path: (String) Path to the csv file containing raw tick data in the format[date_time, price, volume]
+        :param metric: (String) type of imbalance bar to create. Example: dollar_imbalance.
+        :param exp_num_ticks_init: Initial number of expected ticks
+        :param num_prev_bars: Number of previous bars to use in calculations
+        :param num_ticks_ewma_window: Window size for E[T]
+        :param batch_size: Number of rows to read in from the csv, per batch.
+        """
         InformationBars.__init__(self, file_path, metric, exp_num_ticks_init, num_prev_bars, num_ticks_ewma_window,
                                  batch_size)
 
         # Extract bars properties
         self.cache_tuple = namedtuple('CacheData',
-                                      ['date_time', 'price', 'high', 'low', 'tick_rule', 'cum_volume',
-                                       'cum_dollar_value',
+                                      ['date_time', 'price', 'high', 'low', 'tick_rule',
                                        'cum_ticks', 'cum_theta_buy', 'cum_theta_sell', 'exp_num_ticks',
                                        'imbalance_array'])
 
     def _update_counters(self):
+        """
+        Updates the counters by resetting them or making use of the cache to update them based on a previous batch.
+
+        :return: Updated cum_ticks, cum_theta_buy, cum_theta_sell, high_price, low_price, exp_num_ticks, imbalance_array
+        """
         # Check flag
         if self.flag and self.cache:
+            latest_row = self.cache[-1]
             # Update variables based on cache
-            cum_ticks = int(self.cache[-1].cum_ticks)
-            cum_dollar_value = np.float(self.cache[-1].cum_dollar_value)
-            cum_volume = self.cache[-1].cum_volume
-            low_price = np.float(self.cache[-1].low)
-            high_price = np.float(self.cache[-1].high)
-            # cumulative buy and sell imbalances for a particular run calculation (theta_t in Prado book)
-            cum_theta_buy = np.float(self.cache[-1].cum_theta_buy)
-            cum_theta_sell = np.float(self.cache[-1].cum_theta_sell)
-            # expected number of ticks extracted from prev bars
-            exp_num_ticks = np.float(self.cache[-1].exp_num_ticks)
-            # array of latest imbalances
-            imbalance_array = self.cache[-1].imbalance_array
+            cum_ticks = int(latest_row.cum_ticks)
+            low_price = np.float(latest_row.low)
+            high_price = np.float(latest_row.high)
+            # Cumulative buy and sell imbalances for a particular run calculation (theta_t in Prado book)
+            cum_theta_buy = np.float(latest_row.cum_theta_buy)
+            cum_theta_sell = np.float(latest_row.cum_theta_sell)
+            # Expected number of ticks extracted from prev bars
+            exp_num_ticks = np.float(latest_row.exp_num_ticks)
+            # Array of latest imbalances
+            imbalance_array = latest_row.imbalance_array
         else:
             # Reset counters
-            cum_ticks, cum_dollar_value, cum_volume, cum_theta_buy, cum_theta_sell = 0, 0, 0, 0, 0
+            cum_ticks, cum_theta_buy, cum_theta_sell = 0, 0, 0
             high_price, low_price = -np.inf, np.inf
             exp_num_ticks = self.exp_num_ticks_init
             # In run bars we need to track both buy and sell imbalance
             imbalance_array = {'buy': [], 'sell': []}
 
-        return cum_ticks, cum_dollar_value, cum_volume, cum_theta_buy, cum_theta_sell, high_price, low_price, exp_num_ticks, imbalance_array
+        return cum_ticks, cum_theta_buy, cum_theta_sell, high_price, low_price, exp_num_ticks, imbalance_array
 
     def _extract_bars(self, data):
+        """
+        For loop which compiles the various run bars: dollar, volume, or tick.
 
-        cum_ticks, cum_dollar_value, cum_volume, cum_theta_buy, cum_theta_sell, high_price, low_price, exp_num_ticks, imbalance_array = self._update_counters()
+        :param data: (DataFrame) Contains 3 columns - date_time, price, and volume.
+        :return: (List) of bars built using the current batch.
+        """
+        cum_ticks, cum_theta_buy, cum_theta_sell, high_price, low_price, \
+            exp_num_ticks, imbalance_array = self._update_counters()
 
         # Set the first tick rule with 0
         prev_tick_rule = 0
@@ -93,26 +102,12 @@ class RunBars(InformationBars):
             price = np.float(row[1])
             volume = row[2]
 
-            # Calculations
-            dollar_value = price * volume
-            cum_dollar_value = cum_dollar_value + dollar_value
-            cum_volume = cum_volume + volume
+            # Update high low prices
+            high_price, low_price = self._update_high_low(high_price, low_price, price)
 
             # Imbalance calculations
-            try:
-                tick_diff = price - self.cache[-1].price
-                prev_tick_rule = self.cache[-1].tick_rule
-            except IndexError:
-                tick_diff = 0
-
-            tick_rule = np.sign(tick_diff) if tick_diff != 0 else prev_tick_rule
-
-            if self.metric == 'tick_run':
-                imbalance = tick_rule
-            elif self.metric == 'dollar_run':
-                imbalance = tick_rule * volume * price
-            else:  # volume run
-                imbalance = tick_rule * volume
+            signed_tick, prev_tick_rule = self._apply_tick_rule(price, prev_tick_rule)
+            imbalance = self._get_imbalance(price, signed_tick, volume)
 
             if imbalance > 0:
                 imbalance_array['buy'].append(imbalance)
@@ -124,108 +119,75 @@ class RunBars(InformationBars):
                 imbalance_array['buy'].append(0)
                 cum_theta_sell += abs(imbalance)
 
-            if len(imbalance_array['buy']) < exp_num_ticks:
-                # waiting for array to fill for ewma
-                exp_buy_proportion, exp_sell_proportion = np.nan, np.nan
-            else:
-                # expected imbalance per tick
-                ewma_window = int(exp_num_ticks * self.num_prev_bars)
-                buy_sample = np.array(
-                    imbalance_array['buy'][-ewma_window:], dtype=float)
-                sell_sample = np.array(
-                    imbalance_array['sell'][-ewma_window:], dtype=float)
-                buy_and_sell_imb = sum(buy_sample) + sum(sell_sample)
-                exp_buy_proportion = ewma(
-                    buy_sample, window=ewma_window)[-1] / buy_and_sell_imb
-                exp_sell_proportion = ewma(
-                    sell_sample, window=ewma_window)[-1] / buy_and_sell_imb
-
-            # Check min max
-            if price > high_price:
-                high_price = price
-            if price <= low_price:
-                low_price = price
+            exp_buy_proportion, exp_sell_proportion = self._get_expected_imbalance(exp_num_ticks, imbalance_array)
 
             # Update cache
-            cache_data = self.cache_tuple(date_time, price, high_price, low_price, tick_rule, cum_volume,
-                                          cum_dollar_value,
-                                          cum_ticks, cum_theta_buy, cum_theta_sell, exp_num_ticks, imbalance_array)
-            self.cache.append(cache_data)
+            self._update_cache(date_time, price, low_price, high_price, signed_tick, cum_theta_sell, cum_theta_buy,
+                               cum_ticks, exp_num_ticks, imbalance_array)
 
             # Check expression for possible bar generation
-            if max(cum_theta_buy, cum_theta_sell) > exp_num_ticks * max(exp_buy_proportion,
-                                                                        exp_sell_proportion):  # pylint: disable=eval-used
-                # Create bars
-                open_price = self.cache[0].price
-                high_price = max(high_price, open_price)
-                low_price = min(low_price, open_price)
-                close_price = price
-                self.num_ticks_bar.append(cum_ticks)
-                expected_num_ticks_bar = ewma(
-                    np.array(self.num_ticks_bar[-self.num_ticks_ewma_window:], dtype=float),
-                    self.num_ticks_ewma_window)[
-                    -1]  # expected number of ticks based on formed bars
-                # Update bars & Reset counters
-                list_bars.append([date_time, open_price, high_price, low_price, close_price,
-                                  cum_volume, cum_dollar_value, cum_ticks])
-                cum_ticks, cum_dollar_value, cum_volume, cum_theta_buy, cum_theta_sell = 0, 0, 0, 0, 0
+            max_proportion = max(exp_buy_proportion, exp_sell_proportion)
+            if max(cum_theta_buy, cum_theta_sell) > exp_num_ticks * max_proportion:
+                self._create_bars(date_time, price, high_price, low_price, list_bars, cum_ticks)
+
+                # Expected number of ticks based on formed bars
+                exp_num_ticks = ewma(np.array(self.num_ticks_bar[-self.num_ticks_ewma_window:], dtype=float),
+                                     self.num_ticks_ewma_window)[-1]
+
+                # Reset counters
+                cum_ticks, cum_theta_buy, cum_theta_sell = 0, 0, 0
                 high_price, low_price = -np.inf, np.inf
-                exp_num_ticks = expected_num_ticks_bar
                 self.cache = []
 
             # Update cache after bar generation (exp_num_ticks was changed after bar generation)
-            cache_data = self.cache_tuple(date_time, price, high_price, low_price, tick_rule, cum_volume,
-                                          cum_dollar_value,
-                                          cum_ticks, cum_theta_buy, cum_theta_sell, exp_num_ticks, imbalance_array)
-            self.cache.append(cache_data)
+            self._update_cache(date_time, price, low_price, high_price, signed_tick, cum_theta_sell, cum_theta_buy,
+                               cum_ticks, exp_num_ticks, imbalance_array)
         return list_bars
 
-    @staticmethod
-    def _assert_csv(test_batch):
+    def _update_cache(self, date_time, price, low_price, high_price, signed_tick, cum_theta_sell, cum_theta_buy,
+                      cum_ticks, exp_num_ticks, imbalance_array):
         """
-        Tests that the csv file read has the format: date_time, price, and volume.
-        If not then the user needs to create such a file. This format is in place to remove any unwanted overhead.
+        Update the cache which is used to create a continuous flow of bars from one batch to the next.
+
+        :param date_time: Timestamp of the bar
+        :param price: The current price
+        :param low_price: Lowest price in the period
+        :param high_price: Highest price in the period
+        :param signed_tick: The signed tick as defined by the tick rule
+        :param cum_theta_sell: Cumulation of negative signed ticks
+        :param cum_theta_buy: Cumulation of positive signed ticks
+        :param cum_ticks: Cumulative number of ticks
+        :param exp_num_ticks: E{T}
+        :param imbalance_array: (numpy array) of the tick imbalances
         """
-        assert test_batch.shape[1] == 3, 'Must have only 3 columns in csv: date_time, price, & volume.'
-        assert isinstance(test_batch.iloc[0, 1],
-                          float), 'price column in csv not float.'
-        assert isinstance(test_batch.iloc[0, 2],
-                          np.int64), 'volume column in csv not int.'
+        cache_data = self.cache_tuple(date_time, price, high_price, low_price, signed_tick,
+                                      cum_ticks, cum_theta_buy, cum_theta_sell, exp_num_ticks, imbalance_array)
+        self.cache.append(cache_data)
 
-        try:
-            pd.to_datetime(test_batch.iloc[0, 0])
-        except ValueError:
-            print('csv file, column 0, not a date time format:',
-                  test_batch.iloc[0, 0])
+    def _get_expected_imbalance(self, exp_num_ticks, imbalance_array):
+        """
+        Calculate the expected imbalance as defined on page 31 and 32.
 
-    def batch_run(self):
-        # Read in the first row & assert format
-        first_row = pd.read_csv(self.file_path, nrows=1)
-        self._assert_csv(first_row)
-
-        print('Reading data in batches:')
-
-        # Read csv in batches
-        count = 0
-        final_bars = []
-        for batch in pd.read_csv(self.file_path, chunksize=self.batch_size):
-            print('Batch number:', count)
-            list_bars = self._extract_bars(data=batch)
-
-            # Append to bars list
-            final_bars += list_bars
-            count += 1
-
-            # Set flag to True: notify function to use cache
-            self.flag = True
-
-        # Return a DataFrame
-        cols = ['date_time', 'open', 'high', 'low',
-                'close', 'cum_vol', 'cum_dollar', 'cum_ticks']
-
-        bars_df = pd.DataFrame(final_bars, columns=cols)
-        print('Returning bars \n')
-        return bars_df
+        :param exp_num_ticks: Expected number of ticks
+        :param imbalance_array: numpy array of imbalances [buy, sell]
+        :return: expected_buy_proportion and expected_sell_proportion
+        """
+        if len(imbalance_array['buy']) < exp_num_ticks:
+            # waiting for array to fill for ewma
+            exp_buy_proportion, exp_sell_proportion = np.nan, np.nan
+        else:
+            # expected imbalance per tick
+            ewma_window = int(exp_num_ticks * self.num_prev_bars)
+            buy_sample = np.array(
+                imbalance_array['buy'][-ewma_window:], dtype=float)
+            sell_sample = np.array(
+                imbalance_array['sell'][-ewma_window:], dtype=float)
+            buy_and_sell_imb = sum(buy_sample) + sum(sell_sample)
+            exp_buy_proportion = ewma(
+                buy_sample, window=ewma_window)[-1] / buy_and_sell_imb
+            exp_sell_proportion = ewma(
+                sell_sample, window=ewma_window)[-1] / buy_and_sell_imb
+        return exp_buy_proportion, exp_sell_proportion
 
 
 def get_dollar_run_bars(file_path, exp_num_ticks_init, num_prev_bars, num_ticks_ewma_window, batch_size=2e7):
